# openapi.yaml
openapi: 3.0.0
info:
  title: Messages API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get Messages
      description: Retrieves a list of messages.
      parameters:
        - name: status
          in: query
          description: Filter messages by status
          required: false
          schema:
            type: "string"
            enum:
              - "sent"
              - "read"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  messages:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        uuid: 
                          type: "string"
                        text:
                          type: "string"
                        status:
                          type: "string"
                          enum:
                            - "sent"
                            - "read"
              example:
                messages:
                  - uuid: "1f063b8b-a175-6a66-a91a-d58145a4ded4"
                    text: "Hello, World!"
                    status: "read"
                  - uuid: "1f063b8b-a176-693e-943c-d58145a4ded4"
                    text: "How are you today?"
                    status: "sent"
                  - uuid: "1f063b8b-a177-61cc-98b7-d58145a4ded4"
                    text: "Thanks, I'm fine. How about you?"
                    status: "sent"
  /messages/send:
    get:
      summary: Send a Message
      parameters:
        - name: text
          in: query
          description: The text you want to send
          required: true
          schema:
            type: "string"
      responses:
        '204':
          description: Successfully sent
        '400':
          description: Bad request - text parameter is missing or empty
          content:
            text/plain:
              schema:
                type: "string"
              example: "Text is required"